Remove unnecessary type casts

+++ Meter.c
@@ -391,7 +391,7 @@ static void GraphMeterMode_draw(Meter* this, int x, int y, int w) {
    struct timeval now;
    gettimeofday(&now, NULL);
    if (!timercmp(&now, &(data->time), <)) {
-      struct timeval delay = { .tv_sec = (int)(CRT_delay/10), .tv_usec = (CRT_delay-((int)(CRT_delay/10)*10)) * 100000 };
+      struct timeval delay = { .tv_sec = CRT_delay/10, .tv_usec = (CRT_delay-((CRT_delay/10)*10)) * 100000 };
 #ifndef timeradd
       data->time.tv_sec = now.tv_sec + delay.tv_sec;
       data->time.tv_usec = now.tv_usec + delay.tv_usec;
